+++ src/Facades/Shipping.php
@@ -5,6 +5,9 @@ namespace DoubleThreeDigital\SimpleCommerce\Facades;
 use DoubleThreeDigital\SimpleCommerce\Contracts\ShippingManager;
 use Illuminate\Support\Facades\Facade;
 
+/**
+ * @method static \DoubleThreeDigital\SimpleCommerce\Contracts\ShippingManager use($className)
+ */
 class Shipping extends Facade
 {
     protected static function getFacadeAccessor()