+++ tests/Http/Controllers/CouponControllerTest.php
@@ -29,20 +29,24 @@ class CouponControllerTest extends TestCase
     /** @test */
     public function can_store_coupon()
     {
-        $this->markTestSkipped();
-
         Event::fake();
 
         $this->buildCartWithProducts();
 
-        $coupon = Coupon::create([
-            'slug'               => 'half-price',
-            'title'              => 'Half Price',
-            'redeemed'           => 0,
-            'value'              => 50,
-            'type'               => 'percentage',
-            'minimum_cart_value' => null,
-        ])->save();
+        Entry::make()
+            ->collection('coupons')
+            ->id(Stache::generateId())
+            ->slug('half-price')
+            ->data([
+                'title'              => 'Half Price',
+                'redeemed'           => 0,
+                'value'              => 50,
+                'type'               => 'percentage',
+                'minimum_cart_value' => null,
+            ])
+            ->save();
+
+        $coupon = Entry::findBySlug('half-price', 'coupons');
 
         $data = [
             'code' => 'half-price',
@@ -126,6 +130,7 @@ class CouponControllerTest extends TestCase
                 'maximum_uses'       => 5, // We shouldn't be able to use because of this
             ])
             ->save();
+
         $coupon = Entry::findBySlug('half-price', 'coupons');
 
         $data = [
@@ -170,6 +175,89 @@ class CouponControllerTest extends TestCase
     }
 
     /** @test */
+    public function can_store_coupon_limited_to_certain_products_when_product_is_in_cart()
+    {
+        Event::fake();
+
+        $this->buildCartWithProducts();
+
+        Entry::make()
+            ->collection('coupons')
+            ->id(Stache::generateId())
+            ->slug('half-price')
+            ->data([
+                'title'              => 'Half Price',
+                'redeemed'           => 0,
+                'value'              => 50,
+                'type'               => 'percentage',
+                'minimum_cart_value' => null,
+                'products'           => [$this->product->id],
+            ])
+            ->save();
+
+        $coupon = Entry::findBySlug('half-price', 'coupons');
+
+        $data = [
+            'code' => 'half-price',
+        ];
+
+        $response = $this
+            ->from('/cart')
+            ->withSession(['simple-commerce-cart' => $this->cart->id])
+            ->post(route('statamic.simple-commerce.coupon.store'), $data);
+
+        $response->assertRedirect('/cart');
+        $response->assertSessionHasNoErrors();
+
+        $this->cart->find($this->cart->id);
+
+        $this->assertSame($this->cart->data['coupon'], $coupon->id());
+        $this->assertNotSame($this->cart->data['coupon_total'], 0000);
+
+        Event::assertDispatched(CouponRedeemed::class);
+    }
+
+    /** @test */
+    public function cant_store_coupon_limited_to_certain_products_when_products_are_not_in_the_cart()
+    {
+        $this->buildCartWithProducts();
+
+        Entry::make()
+            ->collection('coupons')
+            ->id(Stache::generateId())
+            ->slug('half-price')
+            ->data([
+                'title'              => 'Half Price',
+                'redeemed'           => 5,
+                'value'              => 50,
+                'type'               => 'percentage',
+                'minimum_cart_value' => null,
+                'maximum_uses'       => 0,
+                'products'           => ['another-product-id'],
+            ])
+            ->save();
+
+        $coupon = Entry::findBySlug('half-price', 'coupons');
+
+        $data = [
+            'code' => 'half-price',
+        ];
+
+        $response = $this
+            ->from('/cart')
+            ->withSession(['simple-commerce-cart' => $this->cart->id])
+            ->post(route('statamic.simple-commerce.coupon.store'), $data);
+
+        $response->assertRedirect('/cart');
+        $response->assertSessionHasErrors();
+
+        $this->cart->find($this->cart->id);
+
+        $this->assertNotSame($this->cart->data['coupon'], $coupon->id());
+        $this->assertSame($this->cart->data['coupon_total'], 0000);
+    }
+
+    /** @test */
     public function can_destroy_coupon()
     {
         $this->buildCartWithProducts();
@@ -186,6 +274,7 @@ class CouponControllerTest extends TestCase
                 'minimum_cart_value' => null,
             ])
             ->save();
+
         $coupon = Entry::findBySlug('half-price', 'coupons');
 
         $this->cart->data([
@@ -222,6 +311,7 @@ class CouponControllerTest extends TestCase
                 'minimum_cart_value' => null,
             ])
             ->save();
+
         $coupon = Entry::findBySlug('half-price', 'coupons');
 
         $this->cart->data([